Fix clang-tidy warning: misplaced widening cast in sympathetic resonance test
Fixed bugprone-misplaced-widening-cast at line 1688 by casting start and j
individually before addition instead of casting their sum.

+++ dsp/tests/unit/systems/sympathetic_resonance_test.cpp
@@ -1685,7 +1685,7 @@ TEST_CASE("SympatheticResonance: beating present at 1 Hz separation", "[systems]
     for (int start = analysisStart; start + windowSize < totalSamples; start += hopSize) {
         float peak = 0.0f;
         for (int j = 0; j < windowSize; ++j) {
-            float absVal = std::abs(output[static_cast<size_t>(start + j)]);
+            float absVal = std::abs(output[static_cast<size_t>(start) + static_cast<size_t>(j)]);
             if (absVal > peak) peak = absVal;
         }
         envelope.push_back(peak);
